Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/cached-frame-csp-expected.txt

CONSOLE MESSAGE: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
Check that a CSP header added by a 304 response does not override one from the original request.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that the nonce is the same, meaning that the document was only generated once:
PASS frame1.contentDocument.querySelector("input").value == frame2.contentDocument.querySelector("input").value is true
PASS successfullyParsed is true

TEST COMPLETE
